rag_document/requirements.txt
faiss-cpu
mcp==1.14.1
fastmcp==2.12.4
nltk==3.9.1
numpy==2.3.3
openai==1.109.1
pydantic==2.11.9
python-dotenv==1.1.1
rank-bm25==0.2.2
sentence-transformers==5.1.1
spacy==3.8.7
flask==3.1.2
typing==3.7.4.3
pypdf==4.3.1
python-docx==1.2.0
chardet==5.2.0
striprtf==0.0.26
watchdog==6.0.0
requests==2.32.5
neo4j
fastapi
openpyxl
langchain
langchain-community
chromadb
pandas
groq
tiktoken